Enabling themes to define the UI.
Themes can now overwrite views, layouts, and javascript files.

Note to Reinier: I've implemented this functionality before, but I didn't see it in the code, so I added it again. Let me know if it is now a duplicate.

Added the Kamp Pro 2 theme as a git submodule. I will use this theme to modify the UI to my liking without modifying core.